requirements.txt
pypdf
langdetect
supabase
python-dotenv
lancedb
ollama
pandas
fastapi
uvicorn[standard]
